+++ Model/WebhookEvent.php
@@ -577,7 +577,7 @@ class WebhookEvent extends AbstractModel implements WebhookEventInterface, Ident
                     case 'subtotal':
                         if ($item) {
                             $rawItemPrice = 0.0;
-                            $baseRowItemPrice = 0.0;
+                            $baseRawItemPrice = 0.0;
 
                             $vatPct = 0.0;
 
@@ -595,12 +595,15 @@ class WebhookEvent extends AbstractModel implements WebhookEventInterface, Ident
                             $itemPriceInclTax = 0.0;
                             $baseItemPriceInclTax = 0.0;
 
+                            $itemDiscountAmount = 0.0;
+                            $itemBaseDiscountAmount = 0.0;
+
                             $itemPrice = 0.0;
                             $baseItemPrice = 0.0;
                             foreach ($detail['included'] as $included) {
                                 if ($included['key'] == "item_price") {
                                     $rawItemPrice += $included['price']['amount'];
-                                    $baseRowItemPrice += $included['price']['base']['amount'];
+                                    $baseRawItemPrice += $included['price']['base']['amount'];
 
                                     $itemPrice += $included['price']['amount'];
                                     $baseItemPrice += $included['price']['base']['amount'];
@@ -636,6 +639,13 @@ class WebhookEvent extends AbstractModel implements WebhookEventInterface, Ident
                                     $item->setBaseDiscountAmount($included['price']['base']['amount']);
                                 }
                             }
+
+                            // Split order discount among order items. Note: Order's/Flow's subtotal includes tax.
+                            $itemDiscountAmount += -((($rawItemPrice * $detail['quantity']) /
+                                    ($order->getFlowConnectorItemPrice())) * $order->getDiscountAmount());
+                            $itemBaseDiscountAmount += -((($baseRawItemPrice * $detail['quantity']) /
+                                    ($order->getFlowConnectorBaseItemPrice())) * $order->getBaseDiscountAmount());
+
                             $item->setOriginalPrice($itemPrice);
                             $item->setBaseOriginalPrice($baseItemPrice);
                             $item->setPrice($itemPrice);
@@ -650,13 +660,15 @@ class WebhookEvent extends AbstractModel implements WebhookEventInterface, Ident
                             $item->setRowTotalInclTax($itemPriceInclTax * $detail['quantity']);
                             $item->setBaseRowTotalInclTax($baseItemPriceInclTax * $detail['quantity']);
                             $item->setFlowConnectorItemPrice($rawItemPrice * $detail['quantity']);
-                            $item->setFlowConnectorBaseItemPrice($baseRowItemPrice * $detail['quantity']);
+                            $item->setFlowConnectorBaseItemPrice($baseRawItemPrice * $detail['quantity']);
                             $item->setFlowConnectorVat($vatPrice * $detail['quantity']);
                             $item->setFlowConnectorBaseVat($baseVatPrice * $detail['quantity']);
                             $item->setFlowConnectorDuty($dutyPrice * $detail['quantity']);
                             $item->setFlowConnectorBaseDuty($baseDutyPrice * $detail['quantity']);
                             $item->setFlowConnectorRounding($roundingPrice * $detail['quantity']);
                             $item->setFlowConnectorBaseRounding($baseRoundingPrice * $detail['quantity']);
+                            $item->setDiscountAmount($itemDiscountAmount);
+                            $item->setBaseDiscountAmount($itemBaseDiscountAmount);
                             $item->save();
                         }
                         break;